게시글 목록 조회 필터 변경 및 태그 조회 변경

+++ src/main/java/com/jr_devs/assemblog/repositoryes/JpaPostRepository.java
@@ -22,7 +22,7 @@ public interface JpaPostRepository extends JpaRepository<Post, Long> {
                             @Param("page_size") int pageSize,
                             @Param("order") String order,
                             @Param("order_type") String orderType,
-                            @Param("board_id") int boardId);
+                            @Param("board_id") long boardId);
 
 
     void deleteByTitle(String title);
